+++ src/TokenGenerator/PasswordGenerator.js
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/camelcase */
 import URI from 'urijs';
 import AbstractTokenGenerator from './AbstractTokenGenerator';
 import { memoizePromise } from '../decorator';